`AseCalculation`: add `CMDLINE` key to `settings` node (#12)
This setting allows to define additional command line parameters, such
as adding the `python` executable, which is necessary to run the GPAW
executable. This is explained in the updated `README.md` and a new
example script is added, removing the outdated existing ones.

Finally, a new option is added `optimizer_stdout` which takes the filename
of the file to which the output during a relaxation will be written.