Docs: update the examples in the user guide
The links to the example scripts were broken because they are no longer
stored in the `docs` directory but in the `examples` directory at the
top level of the repository. When building the docs, Sphinx will look
for the files in the same folder as the doc file. Therefore we create a
symlink in the `docs/conf.py` file.

Moreover, the text of the documentation was outdated and has been
updated to reflect the current situation with support for AiiDA v1.0 and
up.